+++ packages/highlight/src/index.d.ts
@@ -27,6 +27,7 @@ export interface RenderHighlightContentProps {
 }
 
 export interface RenderHighlightsProps {
+    highlightAreas: HighlightArea[];
     getCssProperties(area: HighlightArea, rotation: number): React.CSSProperties;
     pageIndex: number;
     rotation: number;
@@ -60,6 +61,7 @@ export enum Trigger {
 // Plugin
 export interface HighlightPlugin extends Plugin {
     jumpToHighlightArea(area: HighlightArea): void;
+    setHighlightAreas(areas: HighlightArea[]): void;
 }
 
 export interface HighlightPluginProps {